Implement a base crud for reusability of code
